@exploitexercise nebula 08
http://www.exploit-exercises.com/nebula/level08 ( http://www.exploit-exercises.com/nebula/level08 )

An administrator captured a login session in a network capture file, which we can replay/reconstruct with tcpdump and tcpflow.

tcpflow performs a console-based equivalent of Wireshark's "follow tcp stream" -- which would have been mildly more useful in this use case because there doesn't appear to be an option to inspect the unprintable characters as hex dumped bytes.