Calling DBus from Emacs
This demo uses DBus from within Emacs. The backend is a ruby server, which is automatically started by DBus whenever it's requested on the session bus. I call one emacs function that delivers messages on the bus and another one that reads messages from the bus.

If you're new to Emacs, keep an eye on the minibuffer at the bottom of my screen.

All the source is in the Git repo ( https://github.com/zmalltalker/deebuzz )